zse-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
apscheduler==3.10.4
pydantic==2.5.0
